requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
motor==2.3.0
multidict==4.7.6
oauth2client==4.1.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.2
pytz==2020.5
requests==2.25.1
rsa==4.6
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
